🔨 [services] Extract function `generate` from `create`
